+++ KaguyaProjectV2/KaguyaBot/Core/Commands/Music/TwitchAudio.cs
@@ -4,6 +4,7 @@ using KaguyaProjectV2.KaguyaBot.Core.Attributes;
 using KaguyaProjectV2.KaguyaBot.Core.Exceptions;
 using KaguyaProjectV2.KaguyaBot.DataStorage.DbData.Queries;
 using System.Threading.Tasks;
+using KaguyaProjectV2.KaguyaBot.Core.Extensions;
 
 namespace KaguyaProjectV2.KaguyaBot.Core.Commands.Music
 {
@@ -24,7 +25,7 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Commands.Music
             var server = await DatabaseQueries.GetOrCreateServerAsync(Context.Guild.Id);
             var user = await DatabaseQueries.GetOrCreateUserAsync(Context.User.Id);
 
-            if (server.IsPremium || user.IsSupporter)
+            if (server.IsPremium || await user.IsPremiumAsync())
             {
                 var playInstance = new Search();
                 var data = await playInstance.SearchAndPlayAsync(Context, query, false, SearchProvider.Twitch);