+++ KaguyaProjectV2/KaguyaBot/Core/Program.cs
@@ -22,7 +22,6 @@ using OsuSharp;
 using System;
 using System.Threading.Tasks;
 using KaguyaProjectV2.KaguyaBot.Core.Handlers.KaguyaPremium;
-using KaguyaProjectV2.KaguyaBot.Core.Handlers.KaguyaSupporter;
 using KaguyaProjectV2.KaguyaBot.Core.Handlers.TopGG;
 using TwitchLib.Api;
 using Victoria;
@@ -188,9 +187,8 @@ namespace KaguyaProjectV2.KaguyaBot.Core
         {
             if (!allShardsLoggedIn) return;
 #if !DEBUG
+            await KaguyaPremiumRoleHandler.Initialize();
             await StatsUpdater.Initialize();
-            await KaguyaSuppRoleHandler.Initialize();
-            await KaguyaSupporterExpirationHandler.Initialize();
             await KaguyaStatsLogger.Initialize();
             await AntiRaidService.Initialize();
             await AutoUnmuteHandler.Initialize();
@@ -200,8 +198,8 @@ namespace KaguyaProjectV2.KaguyaBot.Core
             await KaguyaPremiumExpirationHandler.Initialize();
             await UpvoteExpirationNotifier.Initialize();
             await GameRotationService.Initialize();
-            await ConsoleLogger.LogAsync($"All timers initialized.", LogLvl.INFO);
 #endif
+            await ConsoleLogger.LogAsync($"All timers initialized.", LogLvl.INFO);
         }
 
         private void InitializeEventHandlers()