+++ KaguyaProjectV2/KaguyaBot/DataStorage/DbData/Context/KaguyaDb.cs
@@ -23,7 +23,6 @@ namespace KaguyaProjectV2.KaguyaBot.DataStorage.DbData.Context
         public ITable<Rep> Rep => GetTable<Rep>();
         public ITable<Server> Servers => GetTable<Server>();
         public ITable<ServerExp> ServerExp => GetTable<ServerExp>();
-        public ITable<SupporterKey> SupporterKeys => GetTable<SupporterKey>();
         public ITable<User> Users => GetTable<User>();
         public ITable<WarnSetting> WarnActions => GetTable<WarnSetting>();
         public ITable<WarnedUser> WarnedUsers => GetTable<WarnedUser>();